+++ src/main.js
@@ -2,6 +2,10 @@
  * This file is the first file executed by code using the package
  * Its responsability is to export what is documented
  * Ideally this file should be kept simple to help discovering codebase progressively.
+ *
+ * see also
+ * - https://github.com/davewasmer/devcert
+ * - https://github.com/FiloSottile/mkcert
  */
 
 export {