+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.0.2
+
+- add nssdb store paths on linux
+- improve firefox and chrome detection on linux
+- update dependencies and devDependencies
+- update mac os to 2022 in github workflow
+
 ## 3.0.1
 
 - fix firefox is running detection (could return true because of playwright)